# application/server/openapi.yaml
openapi: 3.0.3
info:
  title: Web Speed Hackathon 2026 API
  version: 1.0.0
servers:
  - url: /api/v1
...
paths:
  /signup:
    post:
      tags:
        - Auth
      summary: サインアップ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
                - name
              properties:
                username:
                  type: string
                password:
                  type: string
                name:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: string
                    enum:
                      - USERNAME_TAKEN
                      - INVALID_USERNAME
  /signin:
    post:
      tags:
        - Auth
      summary: サインイン
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Bad Request
...
  /me:
    get:
      tags:
        - Users
      summary: ログインユーザー情報
      operationId: getCurrentUser
      security:
        - session: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags:
        - Users
      summary: ログインユーザー情報更新
      operationId: updateCurrentUser
      security:
        - session: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{username}:
    get:
      tags:
        - Users
      summary: ユーザー取得
      operationId: getUser
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          $ref: "#/components/responses/NotFound"
  /users/{username}/posts:
    get:
      tags:
        - Users
      summary: ユーザーの投稿一覧
      operationId: getUserPosts
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /dm:
    get:
      tags:
        - DirectMessages
      summary: DM 会話一覧取得
      description: 最新メッセージの新しい順で返し、メッセージが1件もない会話は含みません。
      operationId: getDirectMessageConversations
      security:
        - session: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DirectMessageConversation"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags:
        - DirectMessages
      summary: DM 会話を作成
      operationId: createConversation
      security:
        - session: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - peerId
              properties:
                peerId:
                  type: string
                  format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DirectMessageConversation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /dm/{conversationId}:
    get:
      tags:
        - DirectMessages
      summary: DM 会話取得
      operationId: getConversation
      security:
        - session: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DirectMessageConversation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    head:
      tags:
        - DirectMessages
      summary: DM 会話の状況更新（WebSocket 通信）
      operationId: connectDirectMessageConversation
      security:
        - session: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    required: [type, payload]
                    properties:
                      type:
                        type: string
                        enum: ["dm:conversation:message"]
                      payload:
                        $ref: "#/components/schemas/DirectMessage"
                  - type: object
                    required: [type, payload]
                    properties:
                      type:
                        type: string
                        enum: ["dm:conversation:typing"]
                      payload:
                        type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
  /dm/{conversationId}/messages:
    post:
      tags:
        - DirectMessages
      summary: DM メッセージ送信
      operationId: createDirectMessage
      security:
        - session: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body:
                  type: string
                  minLength: 1
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DirectMessage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /dm/{conversationId}/read:
    post:
      tags:
        - DirectMessages
      summary: DM 既読にする
      operationId: markDirectMessagesRead
      security:
        - session: []
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: connect.sid
      description: express-session によるセッション Cookie
  parameters:
    limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
      description: 取得件数
    offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
      description: オフセット
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
  schemas:
    ProfileImage:
      type: object
      required: [id, alt]
      properties:
        id:
          type: string
          format: uuid
        alt:
          type: string
    User:
      type: object
      required: [id, username, name, description, createdAt, profileImage]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
        profileImage:
          $ref: "#/components/schemas/ProfileImage"
    Image:
      type: object
      required: [id, alt]
      properties:
        id:
          type: string
          format: uuid
        alt:
          type: string
    Movie:
      type: object
      required: [id]
      properties:
        id:
          type: string
          format: uuid
    Sound:
      type: object
      required: [id, title, artist]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        artist:
          type: string
    Post:
      type: object
      required: [id, text, createdAt, user, images]
      properties:
        id:
          type: string
        text:
          type: string
        createdAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
        movie:
          $ref: "#/components/schemas/Movie"
          nullable: true
        sound:
          $ref: "#/components/schemas/Sound"
          nullable: true
        images:
          type: array
          items:
            $ref: "#/components/schemas/Image"
...
    DirectMessageConversation:
      type: object
      required: [id, initiator, member]
      properties:
        id:
          type: string
        initiator:
          $ref: "#/components/schemas/User"
        member:
          $ref: "#/components/schemas/User"
        messages:
          type: array
          items:
            $ref: "#/components/schemas/DirectMessage"
    DirectMessage:
      type: object
      required: [id, body, senderId, conversationId, isRead, createdAt, updatedAt]
      properties:
        id:
          type: string
        body:
          type: string
        senderId:
          type: string
          format: uuid
        conversationId:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        sender:
          $ref: "#/components/schemas/User"